move mail.message model related code
